+++ app/orchestrator/analyzeJob.ts
@@ -1,4 +1,4 @@
-// lib/analyzeJob.ts
+// app/orchestrator/analyzeJob.ts
 export type Tier = "Low" | "Medium" | "High";
 
 export interface RiskResult {
